Move dreamcoder.dreamcoder to dreamcoder.ec and update imports
