+++ bin/regexes.py
@@ -4,7 +4,7 @@ except ModuleNotFoundError:
     import bin.binutil  # alt import if called as module
 
 from dreamcoder.domains.regex.main import main, regex_options
-from dreamcoder.ec import commandlineArguments
+from dreamcoder.dreamcoder import commandlineArguments
 from dreamcoder.utilities import numberOfCPUs
