+++ bin/taskRankGraphs.py
@@ -13,7 +13,7 @@ try:
 except ModuleNotFoundError:
     import bin.binutil  # alt import if called as module
 
-from dreamcoder.ec import *
+from dreamcoder.dreamcoder import *
 import dill
 import numpy as np
 import matplotlib
